trouble
Date and sets are almost implemented, there's a weird bug where if you submit "Lawnmowers" for instance, then you try to submit a different entry for a different day it'll just reinsert the first insert request into the new day. Super weird. I think it's being insertted correctly but its gotta be with the displaying of it.