UI, DB, delete changes
Added a delete button to the view history screen, quick fix in case the user needs to edit a workout for a specific day. They'll have to delete the whole day and relog it. Fixed the WeightSelection screen's display of the current workout being added. UI Improvements to the history and calendar screens. Installing the apk seemed to merge the old database with the new one, not an intended result, need to test its consistency.